:bug: Fix parser stopped to read input after empty line
If the parser found an empty line it stopped reading more
input. The issue was that empty lines and EOF were similar.
As an empty line contains the new line characters (\n, \r\n),
we are using them to distinguish these two situations. The fix
applies to the InputFileDevice and InputConsoleDevice.

Fixes #16